# docs/swagger.yaml
basePath: /
definitions:
...
  model.CustomerUpdatePayload:
    properties:
      name:
        type: string
      password:
        type: string
    type: object
...
  util.Response:
    properties:
      data:
        type: object
      message:
        type: string
    type: object
host: micasvn.ddns.net:9999
info:
  contact: {}
  description: Documentation of Garage Service API
  license: {}
  title: Garage Service API
  version: 1.0.0
paths:
...
  /user/companies:
    get:
      consumes:
      - application/json
      description: Return a list of companies
      parameters:
      - description: Name keyword
        in: query
        name: name
        type: string
      - description: Active state
        in: query
        name: active
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: User API - Company list
      tags:
      - User - Companies
  /user/companies/{id}:
    get:
      consumes:
      - application/json
      description: Return details of a company
      parameters:
      - description: Company's ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: User API - Company detail
      tags:
      - User - Companies
  /user/customers:
    get:
      consumes:
      - application/json
      description: Return details of a customer
      parameters:
      - description: Customer's ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: User API - Customer detail
      tags:
      - User - Customers
    patch:
      consumes:
      - application/json
      description: Update customer's details
      parameters:
      - description: Customer's ID
        in: path
        name: id
        required: true
        type: string
      - description: Customer Update Payload
        in: body
        name: CustomerUpdatePayload
        required: true
        schema:
          $ref: '#/definitions/model.CustomerUpdatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: User API - Customer update
      tags:
      - User - Customers
  /user/services:
    get:
      consumes:
      - application/json
      description: Returns a list of services
      parameters:
      - description: Name keyword
        in: query
        name: name
        type: string
      - description: Company's ID
        in: query
        name: companyid
        type: string
      - description: Active state
        in: query
        name: active
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: User API - List services
      tags:
      - User - Services
  /user/services/{id}:
    get:
      consumes:
      - application/json
      description: Return service's details
      parameters:
      - description: Service's ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/util.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/util.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/util.Response'
      security:
      - BearerToken: []
      summary: User API - Service detail
      tags:
      - User - Services
securityDefinitions:
  BearerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"